added model transformer option for sequence execution repo
