+++ shipyard-controller/db/mongodb_sequence_execution_repo.go
@@ -7,6 +7,7 @@ import (
 	"fmt"
 	apimodels "github.com/keptn/go-utils/pkg/api/models"
 	"github.com/keptn/keptn/shipyard-controller/db/models/sequence_execution"
+	v02 "github.com/keptn/keptn/shipyard-controller/db/models/sequence_execution/v02"
 	"github.com/keptn/keptn/shipyard-controller/models"
 	log "github.com/sirupsen/logrus"
 	"go.mongodb.org/mongo-driver/bson"
@@ -23,12 +24,30 @@ const eventQueueSequenceStateCollectionName = "shipyard-controller-event-queue-s
 var ErrProjectNameMustNotBeEmpty = errors.New("project name must not be empty")
 var ErrSequenceIDMustNotBeEmpty = errors.New("sequence ID must not be empty")
 
-type MongoDBSequenceExecutionRepo struct {
-	DbConnection *MongoDBConnection
+type SequenceExecutionRepoOpt func(repo *MongoDBSequenceExecutionRepo)
+
+func WithSequenceExecutionModelTransformer(transformer sequence_execution.ISequenceExecutionModelTransformer) SequenceExecutionRepoOpt {
+	return func(repo *MongoDBSequenceExecutionRepo) {
+		repo.ModelTransformer = transformer
+	}
 }
 
-func NewMongoDBSequenceExecutionRepo(dbConnection *MongoDBConnection) *MongoDBSequenceExecutionRepo {
-	return &MongoDBSequenceExecutionRepo{DbConnection: dbConnection}
+type MongoDBSequenceExecutionRepo struct {
+	DbConnection     *MongoDBConnection
+	ModelTransformer sequence_execution.ISequenceExecutionModelTransformer
+}
+
+func NewMongoDBSequenceExecutionRepo(dbConnection *MongoDBConnection, opts ...SequenceExecutionRepoOpt) *MongoDBSequenceExecutionRepo {
+	repo := &MongoDBSequenceExecutionRepo{
+		DbConnection: dbConnection,
+		// use the v02 ModelTransformer by default
+		ModelTransformer: &v02.ModelTransformer{},
+	}
+
+	for _, opt := range opts {
+		opt(repo)
+	}
+	return repo
 }
 
 // Get returns all matching sequence executions, based on the given filter
@@ -51,7 +70,7 @@ func (mdbrepo *MongoDBSequenceExecutionRepo) Get(filter models.SequenceExecution
 	for cur.Next(ctx) {
 		var outInterface interface{}
 		err := cur.Decode(&outInterface)
-		sequenceExecution, err := transformBSONToSequenceExecution(outInterface)
+		sequenceExecution, err := mdbrepo.transformBSONToSequenceExecution(outInterface)
 		if err != nil {
 			log.Errorf("Could not decode sequenceExecution: %v", err)
 			continue
@@ -84,7 +103,7 @@ func (mdbrepo *MongoDBSequenceExecutionRepo) GetByTriggeredID(project, triggered
 	if err := item.Decode(&outInterface); err != nil {
 		return nil, err
 	}
-	sequenceExecution, err := transformBSONToSequenceExecution(outInterface)
+	sequenceExecution, err := mdbrepo.transformBSONToSequenceExecution(outInterface)
 	if err != nil {
 		return nil, err
 	}
@@ -116,11 +135,15 @@ func (mdbrepo *MongoDBSequenceExecutionRepo) Upsert(item models.SequenceExecutio
 	}
 	opts := options.Update().SetUpsert(true)
 
-	internalItem := sequence_execution.FromSequenceExecution(item)
-	internalItem.SchemaVersion = sequence_execution.SchemaVersionV02
+	var internalDBItem interface{}
+	if mdbrepo.ModelTransformer != nil {
+		internalDBItem = mdbrepo.ModelTransformer.TransformToDBModel(item)
+	} else {
+		internalDBItem = item
+	}
 
-	filter := bson.D{{"_id", internalItem.ID}}
-	update := bson.D{{"$set", internalItem}}
+	filter := bson.D{{"_id", item.ID}}
+	update := bson.D{{"$set", internalDBItem}}
 
 	_, err = collection.UpdateOne(ctx, filter, update, opts)
 	if err != nil {
@@ -165,7 +188,7 @@ func (mdbrepo *MongoDBSequenceExecutionRepo) AppendTaskEvent(taskSequence models
 	if err != nil {
 		return nil, err
 	}
-	sequenceExecution, err := transformBSONToSequenceExecution(outInterface)
+	sequenceExecution, err := mdbrepo.transformBSONToSequenceExecution(outInterface)
 	if err != nil {
 		return nil, err
 	}
@@ -207,7 +230,7 @@ func (mdbrepo *MongoDBSequenceExecutionRepo) UpdateStatus(taskSequence models.Se
 	if err != nil {
 		return nil, err
 	}
-	sequenceExecution, err := transformBSONToSequenceExecution(outInterface)
+	sequenceExecution, err := mdbrepo.transformBSONToSequenceExecution(outInterface)
 	if err != nil {
 		return nil, err
 	}
@@ -359,28 +382,25 @@ func (mdbrepo *MongoDBSequenceExecutionRepo) updateGlobalSequenceContext(eventSc
 	return err
 }
 
-func transformBSONToSequenceExecution(outInterface interface{}) (*models.SequenceExecution, error) {
+func (mdbrepo *MongoDBSequenceExecutionRepo) transformBSONToSequenceExecution(outInterface interface{}) (*models.SequenceExecution, error) {
 	outInterface, err := flattenRecursively(outInterface)
 	if err != nil {
 		return nil, err
 	}
 
-	data, _ := json.Marshal(outInterface)
+	// if we have a model transformer, use that one to transform the item into a model.SequenceExecution object
+	if mdbrepo.ModelTransformer != nil {
+		return mdbrepo.ModelTransformer.TransformToSequenceExecution(outInterface)
+	}
 
-	internalSequenceExecution := &sequence_execution.JsonStringEncodedSequenceExecution{}
-	if err := json.Unmarshal(data, internalSequenceExecution); err != nil {
+	// if we have no model transformer, directly decode to models.SequenceExecution
+	data, err := json.Marshal(outInterface)
+	if err != nil {
 		return nil, err
 	}
 
-	// if the current schema version is being used, we need to transform it to model.JsonStringEncodedSequenceExecution
-	if internalSequenceExecution.SchemaVersion == sequence_execution.SchemaVersionV02 {
-		transformedSequenceExecution := internalSequenceExecution.ToSequenceExecution()
-		return &transformedSequenceExecution, nil
-	}
-
-	// if the old schema is still being used by that item, we can directly unmarshal it to a model.JsonStringEncodedSequenceExecution
 	sequenceExecution := &models.SequenceExecution{}
-	if err := json.Unmarshal(data, internalSequenceExecution); err != nil {
+	if err := json.Unmarshal(data, sequenceExecution); err != nil {
 		return nil, err
 	}
